# packages/sdk/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Backlog Cloud API
  description: |
    Public REST API for [Backlog Cloud](https://backlog.so). Used by the
    [`backlog`](https://www.npmjs.com/package/backlog) CLI and any
    integration that wants to drive backlog programmatically.

    All endpoints (except `/health`, `/auth/signup`, `/auth/login`,
    and webhook receivers) require a Bearer token in the `Authorization`
    header. Tokens are issued by `/auth/signup` and `/auth/login`,
    expire after 30 days, and use HS256 JWT.
  version: 1.0.0
  license:
    name: Proprietary
    url: https://backlog.so/legal/terms
  contact:
    name: Osmove
    url: https://backlog.so
servers:
  - url: https://backlog.so/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    queryToken:
      type: apiKey
      in: query
      name: user_token
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
        owner_id: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
      required: [id, name, slug, owner_id, created_at]
...
paths:
...
  /projects:
    get:
      summary: List accessible projects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { $ref: '#/components/schemas/Project' }
  /projects/{id}:
    get:
      summary: Show one project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { $ref: '#/components/schemas/Project' }
        '404':
          description: Not found
...